[Security Solution][Detection Engine] Refactor fix-false-positive-alerts skill: extract inline tools, add exception capability, rework content
Refactors the fix-false-positive-alerts Agent Builder skill from a single
937-line monolithic file into a modular inline_tools/ structure following
existing codebase conventions (entity_analytics, security_ml_jobs).

Changes:
- Extract shared helpers into inline_tools/common.ts:
  buildKibanaApiHeaders, getKibanaBaseUrl, runPreview,
  ruleResponseToCreateProps, parseIntervalToMinutes, toBucketList,
  RICH_ALERT_SOURCE_FIELDS, TermsAggBuckets interface
- Split 5 existing inline tools into separate files under inline_tools/
- Add new add-rule-exception tool that creates exception items on a
  rule's default exception list via the Kibana exception list API,
  auto-creating the list and attaching it to the rule if needed
- Rework skill content with full tuning strategy hierarchy:
  1. Rule exceptions (most durable, known-good entities)
  2. Query modification (rule is fundamentally too broad)
  3. Alert suppression (dedup recommendation)
- Update exclusion target guidance to prioritize causal fields:
  parent process > process args > user > host
- Add exception-related best practices to skill guidance

This brings the skill to 6 inline tools (max 7).

Made-with: Cursor
